Update Terraform configurations to use newer provider and chart versions
This commit updates the version constraints for the Kubernetes and Helm providers in `common.tf`, as well as the Helm chart versions for the metrics server, kube-vip, Longhorn, and Prometheus stack. The changes include upgrading the Kubernetes provider to version 2.38, the Helm provider to version 3.0.2, and updating the respective chart versions for kube-vip, Longhorn, and Prometheus to ensure compatibility and access to the latest features and fixes.