Update Terraform configurations to use Helm v3 and improve resource management
This commit updates the Helm provider version to 3.0.2 in the Terraform lock file and modifies the Helm release configurations in `common.tf` and `longhorn.tf` to use the new syntax for setting values. Additionally, the Makefile is enhanced with new cleanup commands for Prometheus-related resources, ensuring a thorough cleanup process. These changes improve compatibility with the latest Helm features and streamline resource management during deployments.